Fixed multiple upload & fetching error
