+++ apps/frontend/app/filter/page.tsx
@@ -22,6 +22,7 @@ const Filter = () => {
 
   const fetchPlaceDetails = async (placeId: string) => {
     const apiKey = process.env.NEXT_PUBLIC_GOOGLE_API_KEY;
+    console.log(apiKey);
     const url = `https://places.googleapis.com/v1/places/${placeId}`;
     const fieldMask = 'id,displayName,photos';
 
@@ -48,7 +49,6 @@ const Filter = () => {
       throw error;
     }
   };
-
 
   const handleSearchStays = async (destination: string) => {
     const apiKey = process.env.NEXT_PUBLIC_GOOGLE_API_KEY;
